Performed cleanup of Hibernate Resources
Our JPA now only uses Eclipselink instead of Hibernate as before. The
main reason of this is because Eclipselink deals better with
@javax.persistence.Id inheritance and GeneratedValue(strategy =
IDENTITY). We really needed this in our entity classes hierarchy.
However, Eclipselink expects weaving, which we disabled for now. At the
end of the project, when Eclipselink will have created the necessary
tables, we may switch back to hibernate if weaving is difficult to
configure with Eclipselink.