Make sure that the web application works in all browsers.
- Before, the web application worked only on Google Chrome. Now, it
works in all browsers. This was made possible by modifying the content
of the templates/administration_layout.xhtml file. In particular, the
fullPage="true" attribute was added to the <p:layout> element. The
templates/administration_layout.xhtml file was renamed to
templates/admin_layout.xhtml to make the name shorter.
- The @Component attribute was added to the
org.rainbow.shopping.cart.ui.web.controller.NavigationController class.
This was to resolve the error (navigationController resolved to null)
thrown by Tomcat.
- By adding the <scope>provided</scope> to the weld-se dependency, it is
now possible for the application to be deployed on Tomcat 8.5. Before,
it worked only on Tomcat v9.0 (TomEE).