Integrate Spring Security with JSF.
Now, it is possible to login into the application. This is because
Spring Security has been integrated with the rainbow shopping cart web
application. However, we noticed that with CSRF (Cross Site Request
Forgery) enabled, the upload of photos with the p:fileUpload component
became impossible. For now, we need to gain time. So we have disabled
CSRF; we will definitely get a workaround and re-enable CSRF in future.