Externalize test database settings and add index to server api app
- The Database class used in the unit tests have been made abstract in
the rainbow-journal-core and rainbow-security-core projects.
- The database connection settings in the MySqlDatabase classes used
during unit tests are now injected so that they match the values in
application.properties files. The values in the later files are those
injected in the datasource properties in the application context files.
- A home page has been added to the rainbow-journal-server project. This
has been made so that the error page displayed after deployment should
no longer be displayed.
- In the rainbow-journal-core and rainbow-security-core projects, the
Cleanup.sql files used to clean up the databases after unit testing have
been adjusted so that the auto_increment on the tables with
auto_increment capabilities should be reset to 1.