Resolve inventory view bug of the rainbow-asset-explorer UI projects
- In the rainbow-asset-explorer UI projects (rainbow-asset-explorer-ui,
rainbow-coronation, and rainbow-etisalat), when a product was added to
the inventory of a location, 
  and the product didn't exist in that inventory before, that product's
number and name were not displayed when the inventory view was opened,
unless the application server 
  was restarted (implying reinitializing the entity manager). That
mearnt that the FetchType.EAGER fetch type introduced on the ManyToOne
annotations of the getProduct() and 
  getLocation() getters of the ProductInventory and ProductAlertInfo
classes in a previous commit did not help. The FetchType.EAGER fetch
type was therefore removed from the 
  later getters, and the below actions carried out:
	1. The ProductInventoryDaoImpl.fixAssociations() method was implemented
to set the associations when creating and updating a ProductInventory
entity.
	2. The InventoryManagerImpl.substract() method was adjusted so that
when products are removed from the inventory, if the remaining quantity
is 0, the corresponding ProductInventory
	   entity should be deleted, instead of continuing to exist with a
quantity of 0. This could be misleading for users if they want to have a
quick glance at the inventory and do not 
	   pay attention to the Quantity column.
	3. The bulk delete operation in InventoryManagerImpl.substract() made
us discover a bug in DaoImpl.delete(List<TEntity> entities) method. In
fact, the line
	   Expression<Boolean> exp =
rt.get(rt.getModel().getId(entityClass)).in(entitiesByIds.keySet());
threw an exception when an embedded ID was used, instead of an ordinary
ID. This line
	   was removed and the method body adjusted. The @Id annotation on the
Identifiable.getId() getter was removed. With this annotation on an
abstract base class, it gave the impression
	   that all derived classes would use a singular ID with this
annotation, which is not true. As can be seen from the ProductInventory
and ProductAlertInfo entities, derived classes of
	   the Identifiable class can use embedded IDs. In a future commit, the
getter Identifiable.getId() and the setter Identifiable.setId() will be
made abstract..