Fix some bugs and refactor entities
- Unnecessary dependencies were removed from the service layer classes.
Such a dependency is the locationDao that existed in the
ProductReceiptServiceImpl class. That dependency was
  completely useless as the class could accomplish its task without it.
Movever, that dependency caused an exception when a quantity in the
ProductReceiptDetail class was updated.
  The bug was because the dependency was not added in the spring
application context XML file.
- ProductReceiptDetailAudit, ProductIssueDetailAudit,
PurchaseOrderDetailAudit, and ShippingOrderDetailAudit have been
adjusted so that the corresponding document ID should not be null
  in the audit tables. This was the case because the document IDs in the
corresponding embeddable were neither insertable nor deletable.
- In the rainbow-asset-explorer-impl,
org.rainbow.asset.explorer.persistence.dao.TrackableDaoImpl has been
deleted. The org.rainbow.orm.listeners.AbstractAuditableEntityListener
and its 
  concrete classes were introduced to play the same role as
org.rainbow.asset.explorer.persistence.dao.TrackableDaoImpl.
- The Auditable annotation has been renamed to AuditableEntity.
- Membership, LoginPolicy, PasswordPolicy and LockoutPolicy have been
made Auditable.
- The entity listener previously used in the rainbow-asset-explorer-orm
project was dependent on the JPA Eclipselink reference implementation.
To make our entity classes unaware on the JPA
  implementation, that listener was moved to a new project, namely,
rainbow-orm-eclipselink. By so doing, we give the client applications
the opportunity to choose which type of listeners to
  use. It is possible that we implement a Hibernate listner in a
different project (may be rainbow-orm-hibernate) in future if we want to
switch to Hibernate.
- Audit classes have been added to the rainbow-security-orm project.