Fix login delete history bug and refactor controllers.
- In the rainbow-security-impl project, in the
PasswordAttemptUpdaterImpl class, the membership_id placeholder used in
the LOGIN_HISTORY_DELETE_QUERY was different to that used 
  in the deleteLoginHistories method. This was due to a type. In fact,
the place holder used in the method was membership_Id. This raised an
SQL exception when the method was
  executed.
- The onToggle method has been moved from DetailController to
ListController.
- In the AbstractDetailController class, the getCurrent() getter has
been renamed to getModel(), and the corresponding setter to setModel().
- The classes with the DetailController suffix have been renamed to now
have the WriteController suffix.
- The classes with the ListController suffix have been renamed to now
have the ReadController suffix.