Fix the immediate user unlock bug
- Before, when a user locked his/her account with wrong password
attempts, he/she could use the unlock user page to attempt to unlock
him/herself with an answer to a recovery question.
  However, when he/she tried to login again, the account was still
locked. If the application server was restarted and the unlock procedure
followed again, the unlock succeeded. That
  meant that the entity manager was guilty. In fact, it was not a nice
idea to implement the methods with UserAccountServiceImpl with JPA. The
unlocking of a user changed information on the 
  membership object, which is a field of the user object. The membership
object could not be saved separately because we were using the UserDao
interface. It could only be committed to the 
  database if and only if the user object was committed. Now, after the
application server was restarted, and the user did three unsuccessful
password attempts, the user object remained in the
  entity manager cache. Even though the membership object changed after
unlocking, calling the UserDao.update method had no effect since the
entity manager saw no change on the user object in cache. 
  The changes were therefore not committed. On the other hand, when the
application server was just restarted and the unlock procedure followed,
no user object was in the cache. So changes were committed
  to the database. In this correction, direct JDBC was used in
UserAccountServiceImpl to correct this bug.