Fix Primefaces theme dependency and fix adjust getLastLoginDate method not to discard time.
- Replaced the primefaces theme dependency groupId from org.primefaces.themes to org.primefaces.extensions,
and changed the version from 1.0.10 to 1.0.8. After cloning the project on my Ubuntu machine, the 1.0.10
dependencies with groupId org.primefaces.themes and version 1.0.10 could no longer be downloaded from the
Maven repository.
- The org.rainbow.security.service.services.UserAccountServiceImpl.getLastLoginDate method has been adjusted
so that the mapRow method now returns rs.getTimestamp("LOGIN_DATE"). Before, it was returning rs.getDate("LOGIN_DATE"),
and was discarding the time part of the last login date.